DomainEnum/requirements.txt
python-whois==0.8.0
colorama==0.4.4
